requirements.txt
pandas>=2.0
numpy>=1.24
yfinance>=0.2.40
matplotlib>=3.8
openpyxl>=3.1
fpdf2>=2.7
scipy>=1.10
